refactor add monitor table into components
